[Issue #2219] Move the location of the foreign and mixin tables
