+++ api/src/db/models/legacy_mixin/forecast_mixin.py
