+++ api/src/db/models/legacy_mixin/opportunity_mixin.py
