+++ api/src/db/models/legacy_mixin/synopsis_mixin.py
