+++ api/src/db/models/staging/synopsis.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Mapped, relationship
 
-from src.db.legacy_mixin import synopsis_mixin
+from src.db.models.legacy_mixin import synopsis_mixin
 from src.db.models.staging.staging_base import StagingBase, StagingParamMixin
 
 from .opportunity import Topportunity
